requirements.txt
torch
ultralytics
Pillow
fastapi
uvicorn